+++ docs/CHANGELOG.md
@@ -23,7 +23,7 @@
 
 ######
 
-# 2025/04/16
+# 2025/04/17
 
 # \_ADDED:
 
@@ -34,8 +34,7 @@
 - Commented out code related to time selector's fixed bar. Visual & behaviour coding will continue later
 - Time Selector: Added OK / Cancel buttons to make saving / cancelling time selection more straightforward
   - Closing Time selector in any method other than clicking OK will result in just closing the popup without saving
-- Time Selector: Now displays currently selected time when opened
-  - WIP: It only displays the DEFAULT TIME atm. Working on fixing it
+- Time Selector: Now displays currently selected time correctly when opened
 
 # 2025/04/15